services/requirements.txt
langchain
langchain-core

langchain-groq

python-dotenv

pdfminer.six

fastapi
uvicorn
python-multipart
